GLK: FROTZ: Further merging together of configuration into a new config.cpp file
