+++ engines/glk/frotz/frotz.cpp
@@ -47,36 +47,6 @@ void Frotz::runGame(Common::SeekableReadStream *gameFile) {
 }
 
 void Frotz::initialize() {
-	if (ConfMan.hasKey("attribute_assignment") && ConfMan.getBool("attribute_assignment"))
-		_attribute_assignment = true;
-	if (ConfMan.hasKey("attribute_testing") && ConfMan.getBool("attribute_testing"))
-		_attribute_testing = true;
-	if (ConfMan.hasKey("ignore_errors") && ConfMan.getBool("ignore_errors"))
-		_ignore_errors = true;
-	if (ConfMan.hasKey("object_movement") && ConfMan.getBool("object_movement"))
-		_object_movement = true;
-	if (ConfMan.hasKey("object_locating") && ConfMan.getBool("object_locating"))
-		_object_locating = true;
-	if (ConfMan.hasKey("piracy") && ConfMan.getBool("piracy"))
-		_piracy = true;
-	if (ConfMan.hasKey("save_quetzal") && ConfMan.getBool("save_quetzal"))
-		_save_quetzal = true;
-	if (ConfMan.hasKey("random_seed"))
-		_random.setSeed(ConfMan.getInt("random_seed"));
-	if (ConfMan.hasKey("script_cols"))
-		_script_cols = ConfMan.getInt("script_cols");
-	if (ConfMan.hasKey("tandy_bit") && ConfMan.getBool("tandy_bit"))
-		_user_tandy_bit = true;
-	if (ConfMan.hasKey("undo_slots"))
-		_undo_slots = ConfMan.getInt("undo_slots");
-	if (ConfMan.hasKey("expand_abbreviations") && ConfMan.getBool("expand_abbreviations"))
-		_expand_abbreviations = true;
-	if (ConfMan.hasKey("err_report_mode")) {
-		_err_report_mode = ConfMan.getInt("err_report_mode");
-		if ((_err_report_mode < ERR_REPORT_NEVER) || (_err_report_mode > ERR_REPORT_FATAL))
-			_err_report_mode = ERR_DEFAULT_REPORT_MODE;
-	}
-
 	// Call process initialization
 	Processor::initialize();
 
@@ -97,7 +67,7 @@ Common::Error Frotz::saveGameData(strid_t file) {
 	if ((gfp = frotzopenprompt(FILE_SAVE)) == nullptr)
 		goto finished;
 
-	if (_save_quetzal) {
+	if (_quetzal) {
 		success = save_quetzal(gfp, story_fp, blorb_ofs);
 	}
 	else {
@@ -164,7 +134,7 @@ Common::Error Frotz::loadGameData(strid_t file) {
 	if ((gfp = frotzopenprompt(FILE_RESTORE)) == nullptr)
 		goto finished;
 
-	if (_save_quetzal) {
+	if (_quetzal) {
 		success = restore_quetzal (gfp, story_fp, blorb_ofs);
 
 	} else {
